Split the long AdminWeb go-file into short files for easy maintanence
